requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
openai==1.6.1
pydantic==2.5.2
python-dotenv==1.0.0